src/third_party/blink/web_tests/http/tests/origin_trials/webexposed/cookiestore-origin-trial-interfaces.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 CookieStore --expire-timestamp=2000000000
-- -->
<meta http-equiv="origin-trial" content="A/tS990j7n9VjcZbX16/nUmZ8VdMMUzvcPOC6lSkRjLCqhbITiPki31a3TpeDuSZzcSKeZ7bik7JB0c+M/i6UQQAAABTeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiQ29va2llU3RvcmUiLCAiZXhwaXJ5IjogMjAwMDAwMDAwMH0=" />
<title>Cookie Store API - interfaces exposed by origin trial</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/origin-trials-helper.js"></script>
<script src="/serviceworker/resources/test-helpers.js"></script>
<script>
test(t => {
  OriginTrialsHelper.check_properties_exist(this, {
     'CookieStore': ['get', 'getAll', 'set', 'delete', 'onchange'],
     'CookieStoreManager': ['getSubscriptions', 'subscribe', 'unsubscribe'],
     'CookieChangeEvent': ['changed', 'deleted'],
     'ServiceWorkerRegistration': ['cookies'],
  });
}, 'Cookie Store API interfaces and properties in Origin-Trial enabled document.');

test(t => {
  assert_true('cookieStore' in self, 'cookieStore property exists on global scope');
  assert_not_equals(window.cookieStore, undefined, 'cookieStore property can be accessed on window');
}, 'Cookie Store API entry point in Origin-Trial enabled document.');

service_worker_test('resources/cookiestore-origin-trial-interfaces-serviceworker-enabled.php');

</script>
